ai-content-factory/requirements.txt
moviepy
gtts
pillow
pyyaml
tqdm
diffusers
transformers
accelerate
torch
opencv-python
google-api-python-client
